+++ backend/app/db/CreditCard.py
@@ -5,5 +5,5 @@ class CreditCard(db.Model):
     number = db.Column(db.String(50))
     expirationDate = db.Column(db.String(20), nullable=False)
     csc = db.Column(db.Integer, nullable=False)
-    balance = db.Column(db.Integer)
+    balance = db.Column(db.Float)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
